Fix postrotate script in /etc/logrotate.d/nginx
We rename nginx to `nginx.real`, and the logrotate script needed to
change to match. We can't use `sv 1 nginx.real` because there isn't a
service called `nginx.real`, so instead we just invoke a `killall`.

This apparently wasn't a problem in previous versions of
passenger-docker because the postrotate script in those builds invoked
`passenger-config reopen-logs`. It's only been since our upgrade to
passenger 1.0.0 that we've seen this problem.